refactor: rename Headingtxt to HeadingText and update imports for consistency
