+++ src/components/common/index.ts
@@ -2,7 +2,6 @@
 export { default as Header } from "./Header.tsx";
 export { default as Footer } from "./Footer.tsx";
 export { default as Img } from "./Image.tsx";
-export { default as Heading } from "./Headingtxt.tsx";
 export { default as Input } from "./Input.tsx";
 export { default as Button } from "./Button.tsx";
 export { default as LoadingSkeleton } from "./LoadingSkeleton.tsx";
@@ -12,3 +11,4 @@ export { default as Pagination } from "./Pagination.tsx";
 export { default as Breadcrumb } from "./Breadcrumb.tsx";
 export { default as Loading } from "./Loading.tsx";
 export { default as Layout } from "./Layout.tsx";
+export * from "./HeadingText.tsx";
