+++ src/components/detailpage/ProductInfo.tsx
@@ -1,4 +1,4 @@
-import Heading from "../common/Headingtxt.tsx";
+import Heading from "../common/HeadingText.tsx";
 import styles from "../../styles/components/detailpage/ProductInfo.module.scss";
 import { useState } from "react";
 import { Icons } from "../../utils/Icons.tsx";
