+++ src/components/filters/FilterSortBar.tsx
@@ -1,7 +1,7 @@
 import styles from "../../styles/components/filters/FilterSortBar.module.css";
 import { Icons } from "../../utils/Icons.tsx";
 import Button from "../common/Button.tsx";
-import Heading from "../common/Headingtxt.tsx";
+import Heading from "../common/HeadingText.tsx";
 import { FilterOverlay } from "./FilterOverlay.tsx";
 import { useState } from "react";
